requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.4.2
umap-learn==0.5.6
matplotlib==3.8.4
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.2.0
